Additional content and routing (#6)
Update content and routing as follows:

HH - E/W - type-of-flat - add new answer option after 'Part of a converted or shared house'. Label to be 'Part of another converted building' and answer guidance to be 'For example, former school, church or warehouse' . No routing or validation changes
HH - E/W - number-of-vehicles - change answer option '4 or more' to be '5 or more'. Add new answer option above with label '4'. No changes to routing or validation
HH/HI - E/W - term-time-address-country - if unanswered then routing should be to end of section
HH/HI - E/W - apprenticeship - include this question in the routing to other-qualifications. Routing should be to other-qualifications if answers to apprenticeship, degree, nvq-level, a-level and gcse are no or not answered
HH/HI - E/W - armed-forces. Change to mutually exclusive answer type.